requirements.txt
streamlit==1.32.0
crewai==0.118.0
crewai-tools==0.42.2
openai==1.62.0
python-dotenv==1.0.1
langchain==0.1.12
langchain-community>=0.0.28 
python-dotenv>=1.1.0
pysqlite3-binary